# api.yaml
openapi: 3.0.2
info:
  title: DBS Programmierpraktikum - Phase 4 - API-Spezifikation
  description: >-
    In der folgenden Liste finden Sie die <b>für die Anwendergruppen
    spezifischen Grundfunktionalitäten</b>. Es müssen <b>alle folgenden
    allgemeinen Bedingungen</b> eingehalten werden: <ul>
      <li><font color="red"><b>Wichtiger allgemeiner Hinweis:</b> Beachten Sie unbedingt die Benennung der Form-Parameter, Query-Parameter sowie die der Ergebnistupel. Bei Abweichungen von in der API angegebenen Benennungen wird der entsprechende Endpunkt als <u><b>falsch</b></u> bewertet.</font></li>
      <li><b>Bei einer gültigen <code>GET</code>-Anfrage wird ein JSON-Array von JSON-Objekten, wobei ein JSON-Objekt eine Entität darstellt, zurückgegeben</b> und dies mit einem Code <code>200</code> belegt. Wird die <code>GET</code>-Anfrage auf einer nicht vorhandenen Ressource ausgeführt, so wird mit Code <code>404</code> geantwortet. Ist die Ressource woanders zu finden, wird dies mit dem Code <code>301</code> und dem Link im Header belegt. Eine <code>GET</code>-Anfrage auf einer Ressource, die eine Sammlung von Entitäten darstellt, kann durch Übergabe von bestimmten Parametern und dazugehörigen Werten <b>case-insensitive gefiltert</b> werden. Dabei muss jede Entität in der Sammlung allen Parametern gleichzeitig genügen. Die Sammlung wird in der Antwort in ein JSON-Array gekapselt.</li>
...
      <li><b>Passwörter werden der Einfachheit halber im Klartext übergeben.</b></li>
...
servers:
...
paths:
  /nutzer:
    get:
      tags:
        - Anwender
      summary: Gibt Nutzer zurück.
      description: 
        Alle **Nutzer** werden zurückgegeben. Falls Query-Parameter mit Werten
        angegeben wurden, wird die Rückgabe noch entsprechend gefiltert.
      parameters:
        - name: email
          in: query
          description: 'Nur Nutzer, deren **E-Mail-Adresse** `email` enthält, werden zurückgegeben.'
          schema:
            type: string
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        nutzerid:
                          type: integer
                          description: '**RowID des Nutzers**.'
                      required:
                        - nutzerid
                    - $ref: '#/components/schemas/nutzer'
    post:
      tags:
        - Anwender
      summary: Fügt einen Nutzer hinzu.
      description:
        Ein **Nutzer** wird mit den entsprechenden Werten hinzugefügt.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/nutzer'
      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'
  /fans:
    get:
      tags:
        - Anwender
      summary: Gibt Fans zurück.
      description: 
        Alle **Fans** werden zurückgegeben. Falls Query-Parameter mit Werten
        angegeben wurden, wird die Rückgabe noch entsprechend gefiltert.
      parameters:
        - name: username
          in: query
          description: 'Nur Fans, deren **Username** `username` enthält, werden zurückgegeben.'
          schema:
            type: string
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        nutzerid:
                          type: integer
                          description: '**RowID des Nutzers**.'
                        fanid:
                          type: integer
                          description: '**RowID des Fan**.'
                      required:
                        - nutzerid
                        - fanid
                    - $ref: '#/components/schemas/nutzer'
                    - $ref: '#/components/schemas/fan'
    post:
      tags:
        - Anwender
      summary: Fügt einen Fan hinzu.
      description:
        Ein **Fan** wird mit den entsprechenden Werten hinzugefügt.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/nutzer'
                - $ref: '#/components/schemas/fan'
      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'
  /mitarbeiter:
    get:
      tags:
        - Anwender
      summary: Gibt Mitarbeiter zurück.
      description: 
        Alle **Mitarbeiter** werden zurückgegeben. Falls Query-Parameter mit Werten
        angegeben wurden, wird die Rückgabe noch entsprechend gefiltert.
      parameters:
        - name: personalnummer
          in: query
          description: 'Nur Mitarbeiter, deren **Personalnummer** == `personalnummer` ist, werden zurückgegeben.'
          schema:
            type: integer
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        nutzerid:
                          type: integer
                          description: '**RowID des Nutzers**.'
                        mitarbeiterid:
                          type: integer
                          description: '**RowID des Mitarbeiters**.'
                      required:
                        - nutzerid
                        - mitarbeiterid
                    - $ref: '#/components/schemas/nutzer'
                    - $ref: '#/components/schemas/mitarbeiter'
    post:
      tags:
        - Anwender
      summary: Fügt einen Mitarbeiter hinzu.
      description:
        Ein **Mitarbeiter** wird mit den entsprechenden Werten hinzugefügt.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/nutzer'
                - $ref: '#/components/schemas/mitarbeiter'
      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'
...
  /fans/{fanid}/mannschaft:
    get:
      tags:
        - Anwender
      summary: Gibt die vom Fan gefolgte Mannschaft zurück.
      description: 
        Die vom Fan **gefolgte Mannschaft** wird zurückgegeben.
      parameters:
        - name: fanid
          in: path
          description: '**RowID** des Fans.'
          schema:
           type: integer
          required: true
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        mannschaftid:
                          type: integer
                          description: '**RowID** der Mannschaft.'
                      required:
                        - mannschaftid
        '404':
          $ref: '#/components/responses/404'
    post:
      tags:
        - Authentifizierter und als Fan authorisierter Anwender
      summary: Eine Mannschaft wird von einem Fan gefolgt.
      security:
        - basicAuth: []
      description: Die gewählte Mannschaft mit **MannschaftID** wird von dem authentifizierten Fan mit **FanID** gefolgt.
      parameters:
      - name: fanid
        in: path
        description: '**RowID** des Fans.'
        schema:
          type: integer
        required: true
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - type: object
                  properties:
                    mannschaftid:
                      type: integer
                      description: '**RowID** der Mannschaft.'
                  required:
                    - mannschaftid
      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
    patch:
      tags:
        - Authentifizierter und als Fan authorisierter Anwender
      security:
        - basicAuth: []
      summary: Ändert die gefolgte Mannschaft.
      description: 
        Die **gefolgte Mannschaft** wird durch eine **neue Mannschaft** ersetzt.
      parameters:
        - name: fanid
          in: path
          description: '**RowID** des Fans.'
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - type: object
                  properties:
                    mannschaftid:
                      type: integer
                      description: '**RowID** der neuen Mannschaft.'
                  required:
                    - mannschaftid
      responses:
        '204':
          $ref: '#/components/responses/204'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  responses:
    '201':
      description: '**Created**.'
      headers:
        Location:
          description: Die URL der neuen Ressource. Siehe hierzu auch die Anmerkung bzgl. fiktiver Links im ersten Abschnitt.
          schema:
            type: string
    '204':
      description: '**No Content**.'
...
    '400':
      description: '**Bad Request**.'
    '401':
      description: '**Unauthorized**.'
      headers:
        WWW-Authenticate:
          schema:
            type: string
    '403':
      description: '**Forbidden**.'
    '404':
      description: '**Not Found**.'
  schemas:
    nutzer:
      type: object
      properties:
        email:
          type: string
          description: '**E-Mail-Adresse** des Nutzers.'
        passwort:
          type: string
          description: '**Passwort** des Nutzers.'
      required:
        - email
        - passwort
    fan:
      type: object
      properties:
        username:
          type: string
          description: '**Username** des Fans.'
      required:
        - username
...
    mitarbeiter:
      type: object
      properties:
        personalnummer:
          type: integer
          description: '**Personalnummer** des Mitarbeiters.'
      required:
        - personalnummer
...